requirements.txt
### ✅ `requirements.txt`

```txt
# Web scraping
playwright==1.44.0
playwright-stealth==1.0.6

# Web application
flask==2.3.3
flask-cors==4.0.0

# Database
pymongo==4.6.1

# Google OAuth & Sheets
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.128.0
oauth2client==4.1.3  # optional, if used for legacy gspread

# Excel / CSV
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1

# Environment & Logging
python-dotenv==1.0.1
rich==13.7.1

# Emailing
yagmail==0.15.293  # Optional, if you're using it
```

---

### ⚙️ After creating this file:

Install everything in one go:

```bash
pip install -r requirements.txt
playwright install
```

This ensures:

* All necessary Python packages are installed.
* Playwright's browser binaries (Chromium, Firefox, WebKit) are available.
